refactor: update aboutsopt
